test(TASK-07): split my component PokemonDetailPageCard to an another proper file
